Fix most bugs

+++ src/EasyNuitka.py
@@ -2,10 +2,10 @@
 import os
 import sys
 import threading
-
 from PyQt5.QtCore import Qt
 from PyQt5 import QtWidgets,QtGui
-from PyQt5.QtWidgets import QApplication
+from PyQt5.QtGui import QIcon
+from PyQt5.QtWidgets import QApplication, QStyleFactory
 import nuitka
 import mainui
 import subprocess
@@ -17,42 +17,62 @@ class MainGui(mainui.Ui_MainWindow,	QtWidgets.QMainWindow):
 		super(MainGui, self).__init__(parent)
 		self.setupUi(self)
 		self.setWindowTitle( "EasyNuitka")
+		self.setWindowIcon(QIcon("./icon/Nuitka.png"))
 		self.pushButton.clicked.connect(self.choose_main_py)
 		self.pushButton_2.clicked.connect(self.set_output_path)
 		self.pushButton_7.clicked.connect(self.Multithreading_command)
-		self.pushButton_4.clicked.connect(self.nofollow_import_to)
-		self.pushButton_12.clicked.connect(self.nofollow_import_to_remove)
+		self.pushButton_4.clicked.connect(self.follow_import_to)
+		self.pushButton_12.clicked.connect(self.follow_import_to_remove)
 		self.pushButton_20.clicked.connect(self.set_icon)
 		self.pushButton_19.clicked.connect(self.remove_icon)
+		self.pushButton_9.clicked.connect(self.quit)
+		self.pushButton_3.clicked.connect(self.nofollow_import_to)
+		self.pushButton_10.clicked.connect(self.nofollow_import_to_remove)
+		self.pushButton_17.clicked.connect(self.include_package_remove)
 		self.chekbox_clicked_init()
 		self.command_dict={}
 		self.command_str=None
 		self.main_name=None
 		self.outputpath="./"
+		self.follow_import_to_document_str=""
 		self.nofollow_import_to_document_str=""
-		self.iconname=""
-
+		self.iconname=None
+		self.iconpath=""
+		self.compile_module_path=None
+		self.statusbar.showMessage("The software runs normally")
 	def choose_main_py(self):
 		try:
-			file = QtWidgets.QFileDialog.getOpenFileName(self,
-														 "getOpenFileName", "./",
-														 "All Files (*.py);;Text Files (*.py)")
-			self.filenpath=file[0]
-			self.lineEdit.setText(file[0])
-			self.main_name=self.lineEdit.setText(file[0])
-			self.filename=self.filenpath.split("/")[-1]
-			self.filenpath=self.filenpath.replace(self.filename,"")
-			self.dish=self.filenpath[0:2]
+			if not self.checkBox_7.isChecked():
+				file = QtWidgets.QFileDialog.getOpenFileName(self,
+															 "choose main.py", "./",
+															 "All Files (*.py);;Text Files (*.py)")
+				self.filenpath = file[0]
+				self.lineEdit.setText(file[0])
+				self.main_name = self.lineEdit.setText(file[0])
+				self.filename = self.filenpath.split("/")[-1]
+				self.filenpath = self.filenpath.replace(self.filename, "")
+				self.dish = self.filenpath[0:2]
+				self.statusbar.showMessage("File selected successfully")
+
+			else :
+				directory = QtWidgets.QFileDialog.getExistingDirectory(self, "choose compile module dir", "./")
+				self.compile_module_path=directory
+				self.lineEdit.setText(directory)
+				self.lineEdit_9.setText(self.compile_module_path.split("/")[-1])
+				self.statusbar.showMessage("Output directory set successfully")
+
 
 		except Exception as e:
+			self.statusbar.showMessage("File selection failed")
 			print(e)
 	def set_output_path(self):
 		try:
 			directory = QtWidgets.QFileDialog.getExistingDirectory(self, "getExistingDirectory", "./")
 			self.outputpath=directory
 			self.lineEdit_2.setText(directory)
-
+			self.statusbar.showMessage("Output directory set successfully")
 		except Exception as e:
+			self.statusbar.showMessage("Output directory setting failed")
 			pass
 	def get_package_method(self):
 		#--follow-imports
@@ -82,23 +102,29 @@ class MainGui(mainui.Ui_MainWindow,	QtWidgets.QMainWindow):
 			self.checkBox_8.setEnabled(False)
 			self.checkBox_3.setEnabled(False)
 			self.checkBox_4.setEnabled(False)
-			self.checkBox_9.setEnabled(False)
-			self.checkBox_17.setEnabled(False)
 			self.lineEdit_3.setEnabled(False)
 			self.lineEdit_4.setEnabled(False)
 			self.lineEdit_5.setEnabled(False)
 			self.lineEdit_8.setEnabled(False)
+			self.lineEdit_6.setEnabled(False)
+			self.lineEdit_9.setEnabled(True)
+			self.radioButton.setEnabled(False)
+			self.radioButton_2.setEnabled(False)
 			self.command_dict[self.checkBox_7] = "--module"
 			#print(self.command_dict)
 		elif not self.checkBox_7.isChecked():
 			self.checkBox_3.setEnabled(True)
 			self.checkBox_4.setEnabled(True)
-			self.checkBox_9.setEnabled(True)
-			self.checkBox_17.setEnabled(True)
 			self.lineEdit_3.setEnabled(True)
 			self.lineEdit_4.setEnabled(True)
 			self.lineEdit_5.setEnabled(True)
 			self.lineEdit_8.setEnabled(True)
+			self.lineEdit_6.setEnabled(True)
+			self.lineEdit_9.setEnabled(False)
+			self.radioButton.setEnabled(True)
+			self.radioButton_2.setEnabled(True)
+			self.checkBox_3.setEnabled(True)
+			self.checkBox_4.setEnabled(True)
 			if self.checkBox.isChecked():
 				self.checkBox_8.setEnabled(False)
 			if self.checkBox_8.isChecked():
@@ -110,19 +136,21 @@ class MainGui(mainui.Ui_MainWindow,	QtWidgets.QMainWindow):
 			self.checkBox_4.setEnabled(False)
 			self.checkBox_7.setEnabled(False)
 			self.command_dict[self.checkBox_3] = "--standalone"
-		else:
-			self.checkBox_4.setEnabled(True)
+		elif not self.checkBox_3.isChecked() and not self.checkBox_7.isChecked():
 			self.checkBox_7.setEnabled(True)
+			self.checkBox_4.setEnabled(True)
 			self.command_dict[self.checkBox_3] = ""
 		#--onefile
 		if self.checkBox_4.isChecked():
 			self.checkBox_3.setEnabled(False)
 			self.checkBox_7.setEnabled(False)
 			self.command_dict[self.checkBox_4] = "--onefile"
-		else:
+		elif not self.checkBox_4.isChecked() and not self.checkBox_7.isChecked() :
 			self.checkBox_3.setEnabled(True)
-			self.checkBox_7.setEnabled(True)
+			if not self.checkBox_3.isChecked():
+				self.checkBox_7.setEnabled(True)
 			self.command_dict[self.checkBox_4] = ""
+
 		#--mingw64
 		if self.checkBox_2.isChecked():
 			self.checkBox_10.setEnabled(False)
@@ -222,28 +250,43 @@ class MainGui(mainui.Ui_MainWindow,	QtWidgets.QMainWindow):
 			self.command_dict[self.radioButton] = "--windows-disable-console"
 		if self.radioButton_2.isChecked():
 			self.command_dict[self.radioButton_2] = ""
-
+		self.statusbar.showMessage("Selection succeeded")
 	def Multithreading_command(self):
 		t=threading.Thread(target=self.excute_command,args=())
 		t.start()
+		t.join()
+		self.statusbar.showMessage("packaging is finishing")
 	def excute_command(self):
 		try:
 			self.command_str="python -m nuitka "
 			for command in self.command_dict.keys():
 				if self.command_dict[command]=="":
 					continue
+				if self.command_dict[command]=="--module":
+					self.command_dict[command]="--module "+self.compile_module_path.split("/")[-1]+" "+"--include-package="+\
+											   self.lineEdit_9.text()
 				self.command_str+=self.command_dict[command]+" "
-
+			if self.follow_import_to_document_str!="":
+				self.command_str += "--follow-import-to=" + self.follow_import_to_document_str[0:len(self.follow_import_to_document_str)-1] + " "
+			if self.follow_import_to_document_str!="":
+				self.command_str+="--nofollow-import-to="+self.follow_import_to_document_str+" "
 			self.command_str+="--output-dir="+self.outputpath+" "
-			self.command_str+="--windows-icon-from-ico="+self.iocnname+" "
-			self.command_str+=self.filename
+			if self.iconname!=None:
+				self.command_str+="--windows-icon-from-ico="+self.iconname+" "
+			if not self.checkBox_7.isChecked():
+				self.command_str+=self.filename
 
 			print(self.command_str)
-			os.chdir(self.filenpath)
+			if not self.checkBox_7.isChecked():
+				os.chdir(self.filenpath)
+			else:
+				os.chdir(self.compile_module_path)
+				os.chdir("../")
 			os.system(self.command_str)
-
+			self.statusbar.showMessage("Please wait to start packaging......")
 		except Exception as e:
-			print(e)
+			self.statusbar.showMessage(e)
+			#print(e)
 
 	def chekbox_clicked_init(self):
 		self.package_method=[]
@@ -266,51 +309,69 @@ class MainGui(mainui.Ui_MainWindow,	QtWidgets.QMainWindow):
 				check.clicked.connect(self.get_package_method)
 			except:
 				pass
-
+		self.lineEdit_9.setEnabled(False)
 	def nofollow_import_to(self):
+		self.follow_import_to_document_str=self.lineEdit_3.text()
+		self.statusbar.showMessage("add module to nofollow_import_to ")
+	def nofollow_import_to_remove(self):
+		self.lineEdit_3.setText("Enter module/package ;Separate them with commas")
+		self.follow_import_to_document_str=""
+		self.statusbar.showMessage("Removed")
+	def follow_import_to(self):
 		try:
 			directory = QtWidgets.QFileDialog.getExistingDirectory(self, "getExistingDirectory", "./")
-			self.nofollow_import_to_document = directory.split("/")[-1]
-			self.nofollow_import_to_document_str+=self.nofollow_import_to_document+","
-			self.lineEdit_4.setText(self.nofollow_import_to_document_str)
+			self.follow_import_to_document = directory.split("/")[-1]
+			self.follow_import_to_document_str+=self.follow_import_to_document+","
+			self.lineEdit_4.setText(self.follow_import_to_document_str)
 		except Exception as e:
 			print(e)
 			pass
-	def nofollow_import_to_remove(self):
+	def follow_import_to_remove(self):
 		try:
-			self.lineEdit_4.setText("")
+			self.lineEdit_4.setText("Select the folder you want to compile")
 		except Exception as e:
-			print(e)
+			self.statusbar.showMessage("Removed")
 			pass
+	def include_data_files(self):
+		pass
 	def set_icon(self):
 		try:
-			file = QtWidgets.QFileDialog.getOpenFileName(self,
-														 "getOpenFileName", "./",
-														 "All Files (*);;Text Files (*)")
-			self.iconpath=file[0]
-			self.lineEdit_6.setText(self.iconpath)
-			src=self.iconpath
-			dist=self.filenpath+src.split("/")[-1]
-			try:
-				shutil.copyfile(src,dist)
-			except:
-				pass
-			finally:
-				self.iocnname = src.split("/")[-1]
-
+			if not self.checkBox_7.isChecked():
+				file = QtWidgets.QFileDialog.getOpenFileName(self,
+															 "getOpenFileName", "./",
+															 "All Files (*);;Text Files (*)")
+				self.iconpath = file[0]
+				self.lineEdit_6.setText(self.iconpath)
+				src = self.iconpath
+				dist = self.filenpath + src.split("/")[-1]
+				if not os.path.exists(dist):
+					shutil.copyfile(src, dist)
+				self.iconname = src.split("/")[-1]
 
 		except Exception as e:
 			print(e)
 	def remove_icon(self):
-		self.lineEdit_6.setText("")
+		self.lineEdit_6.setText("Select  icon")
+		self.statusbar.showMessage("Removed")
 
+	def include_data_files(self):
+		try:
+			file = QtWidgets.QFileDialog.getOpenFileName(self,
+														 "getOpenFileName", "./",
+														 "All Files (*);;Text Files (*)")
+			self.include_data_files=file[0]
+		except Exception as e:
+			pass
 
-
-
+	def include_package_remove(self):
+		self.lineEdit_9.setText("Default to the entire module")
+	def quit(self):
+		sys.exit()
 # Press the green button in the gutter to run the script.
 if __name__ == '__main__':
-	app = QtWidgets.QApplication(sys.argv)
 	QApplication.setAttribute(Qt.AA_EnableHighDpiScaling)
+	app = QtWidgets.QApplication(sys.argv)
+	QApplication.setStyle(QStyleFactory.create('Fusion'))
 	win = MainGui()
 	win.show()
 	sys.exit(app.exec_())
